refactor: media card folder structure
